feat(verify): support main JS coding standards (AirBnB, Google, StandardJS)
BREAKING CHANGE: The 'verify' generator now specifies a **coding style* instead of
a linting engine (except in the case of TypeScript, which is not as mature in this regard).
As there are no well-known CSS coding styles, verify still specifies the CSS linter in terms
of the linting engine.

Close #16